Finished Menu system for tools
Toolbar is a container for all tool widgets, it holds toolboxes like the 'machines' toolbox and some generic toolbuttons that do not fit in categories like the build tool.
Toolboxes are containers for multiple tools. They store a list of tools that are in their category. When clicked, they expand leftwards to show these buttons.